Provide basic descriptions of utils and bootstrap
This change set provides the base line of method descriptions that are
needed to support the functions within the agent.

Supports bringing up the Neutron Agent and listing the current Network
Bridges on the system.

Includes a fix to get the unit tests running properly with tox on the
server. That fix also supports preparing a policy.json file for use
within the Neutron source tree.  Only needed as we are not directly in
the Neutron project.

Change-Id: I759c807551f71104e01b2d2b3cb71f3e4a530bf0