Provide baseline Fake PowerVM Agent
This change provides the initial code that can be enhanced to build a
PowerVM Shared Ethernet Neutron Simulation Agent.  That agent would not
require hardware underneath to run, and would provide a means to scale
out testing on the message bus.

This is just the initial framework.  Enhancements will be needed as the
proper agent is built out.

Change-Id: Ida4a0211de508d588f7c2b0f9cfc0b0a3f47735a