Updating README.rst for stackforge
Updates to the README.rst were done to include latest set of
information.  Main changes are the inclusion of the pypowervm API and
indication that Neutron team wishes for third party agents to be
independent of the core neutron project.

Also updates the properties for communication with the API to reflect
more generic naming.

Includes a tox update to handle running py27 target with the new neutron
updates.

Change-Id: Iffcb24e4eee100eb43665e62e6ceb493d002341c