Provide heal method for VLAN configuration
ML2 agents tend to support a rebuild of the bridges upon a restart of
the agent.  This change set provides a baseline for the agent such that
it rebuilds properly upon boot up.  The clean up should:
 - Ensure that each Neutron Port is served via the Network Bridge
 - Remove any non-required VLANs off the bridges
 - Only removes VLANs that aren't in use, whether it is managed by
   Neutron or externally.

The clean method also runs periodically during query.  It does this as
it is a moderately expensive method call.  Future change sets will
optimize this flow.

Change-Id: I3be631a46adc9bbfc42649fe65ac7f203a0a65f9