Separate ML2 agent framework from logic
There are a few pieces of framework that are required to have the ML2
agent properly communicate with the plugin and Neutron server.  This
change set separates out a baseline set of classes specific PowerVM
agents can use to extend for differening technology bases.

The result is that the SharedEthernetNeutronAgent is focused primarily
on the logic needed to provision against a SharedEthernetAdapter (or
NetworkBridge via the API).

Also changed looping approach to be time based.

Change-Id: Ia4a369d4b20cf084a9a28cf701fe94efaa252a2d