Parse Bridge Mappings to include XAG
The pypowervm library recently defaulted the XAG attribute (extended
attributes) to none.  The side effect to this is that the VIOS read is
no longer returning the Shared Ethernet Adapters.

This change updates the parse of the bridge mappings to pass in the
appropriate XAG to get the proper return data.

It also includes a unit test for the parse_bridge_mappings which had
been missed initially.

Change-Id: Ifb9ce3b9c008cf71715667d22312b8fa6a87304a