Support agent boot without configuration
The configuration requires a bridge_mapping which indicates which
Network Bridge to use for a given Neutron physical network.  Some
deployments may just have a single Network Bridge.  In those cases, it
is an obvious scenario to use single Network Bridge against the default
physical network.

This change set supports not mandating the bridge_mappings to be
specified if the host has a single Network Bridge.  It will log a
warning message as this is not ideal on agent boot up.

This change set also moves exceptions into a common class.  This is an
ancillary change that was done to make unit testing the function easier.

Change-Id: I9ee36e453dd611382c79ff937e84abead16e1e5f