Change updated ports from set to list
The updated ports in the agent base is a list of neutron ports that have
been updated since the last rpc_loop.

It was a set of ports.  However, the port is non-hashable and therefore
can not be added to a set.  This change set moves the container to a
list.

Change-Id: Ic4b65e575efa54d51f0cefa2472f94bfca449675