Add a call back to the update port status
Once a port has been successfully provisioned, the status of the device
in Neutron needs to be updated.  This means that it can not be done
until the Network Bridge is ready for the traffic and the Client Network
Adapter is running on the correct VLAN.

There are a couple hooks to keep in mind.  First, the port is
provisioned as soon as Nova ties it to a host.  This allows us to
immediately plug it into the Network Bridge.  However, the second hook
is to the VIF (in the PowerVM scenario - the Client Network Adapter)
where we must update the PVID.  This is done asynchronously, only able
to be done when Nova creates the VIF.

To properly update the device status, it must not be done until both the
Network Bridge and the Client Network Adapter are updated.

This change set ties the overall provision path to set the status down
if the immediate action (Network Bridge provision) fails.  Then the
PVIDLooper will update the specific device up/down depending on its
result.  This maintains the lifecycle flow and allows for proper
updating of the device status back to Neutron.

Change-Id: I401876b08a4b25fa7211fee6f595ec392313ca5a