Account for pending requests in heal
The Nova process can have a delay between identifying that a neutron
port is on a given host and when it creates the VIF.  The PVIDLooper was
created to handle that delay.

However it has an issue where if the delay was too long, it would time
out and error.  Also, the optimize code had an exposure where it did not
take into account pending VLAN IDs.

The delay from nova could be particularly high if there was BDM's in
use.

This change set resolves both of these issues.  It introduces a timeout
for the pvid update.  This becomes user configurable, but defaults to
the standard VIF timeout.  While these are not 100% the same, it is a
similar concept.

It also incorporates a check into the heal code to take into account
pending VLAN provisions.

Change-Id: I75be9b472dcb9323a8f5a893fbf3442247c1b4d0
Closes-Bug: 1466974