Add retry logic to CNA update
The update of the CNA pvid is critical to getting the VM online.
The feed of CNAs is done up front and then looped over periodically.
However, there may be times when the CNA obtained from the feed is not
the absolute current by the time that the agent attempts to update it.
In these scenarios the API throws a 'etag' error.

This change set adds logic surrounding this to make sure that the API
can get the latest CNA and attempt the update again.

Change-Id: Ie31cb3c81c3bedb112f9866da7d683b780482648