Refactor the utility module
This change set provides a refactoring of the utility module to make it
simple functions instead of the heavier weight class.  The agent itself
now has the adapter, which is in line with the PowerVM code in other
OpenStack projects.

This change set also changes single quotes to double quotes to keep
method comments consisten within the project.

Change-Id: I006b43b8c8a7321f875a169c8285eff6a4d0f12e