Add try/except around PVID Updater looping call
In rare scenarios the PVIDUpdater may throw an exception back from the
API.  This was causing the FixedIntervalLoopingCall to stop processing
new requests.

This change adds a simple try/except (which adds logging) around the
PVIDUpdater's execute method.  This will log the exception, but allow
further requests to go through.

Change-Id: I2fd10d265d957c7ea47e34e3f64fe0d10ac03396
Closes-Bug: 1484909