Improve provision VLAN performance
The VLAN provisioning performance can be sped up by querying only for
the CNAs of the VM that the request is for.  The device_id can be used
for this data - as that matches the nova instance UUID.

This is valuable in scenarios where there are hundreds of VMs on the
system.  Rather than having every request query the entire system, this
change allows us to only get the data that we need for the request that
comes in.

This change required bundling together data from both the neutron port
and the device detail.  As such, a new entity called NetworkDevice was
added that pulls together these two pieces of information.

Change-Id: I0a4a7cd2a2f3681598aec368ddd4d6c0366654b5