Remove fixed interval looper to eventlet thread
This change modifies the PVID Looper to use an eventlet based thread.
This resolves an excessive logging issue that could occur.

The previous mechanism would always run a new pass after 1 second -
regardless of if the previous loop had finished.  This would allow
multiple update requests to go through at once.  Subsequent runs could
actually send a failure request.

This change makes it so that only one loop pass can run at a time.  It
also adds some additional logging in the event of a failure to help
guide the admin to the root issue.

Change-Id: I9d5b0e023f91bbc3817ecbb6b395f5f6230214b7