List LPARs before querying for a given CNA
This change set will list all of the VM UUIDs out during the
UpdateVLANRequest.  It does this so that it does not query for the CNA's
on VMs that do not exist.

This caused issues in the backing PowerVM API due to a tremendous number
of requests for elements that did not exist.  This reduces that load
dramatically as the majority were coming from Neutron.

Change-Id: I36ba4de7cb2b16981cf10f4e9825436ab2240a64