Optimize Provision Requests
This change set resolves two major issues.  First, the agent hadn't been
taking into account that when a provision request comes into the agent,
it hasn't yet been identified to be on that host.  The way the logging
was set up always assumed that the port was targetted for a given host.

In large scale environments, this led to excessive logging when ports
were not there.  Further optimization is needed in this space, but this
solves the immediate logging concern.  It also stops another host from
marking a port down.

Secondly, the new provision request was looking at the wrong attributes,
and this caused the requests to be O(n^2).  This sent many incorrect
down requests for a port.  While it worked in smaller environments,
scaling out didn't work properly.

Change-Id: I1272350382be1039c48e182961a3c6855e9ae232