Increase pvid looping count
Recent large scale environment testing indicates that the response time
from Nova (between when the port is bound to a host, versus actually
existing on the system) can be significant.  This is typically occurring
when a large number of concurrent deploys are occurring - and each has
cinder volumes.

This change increases the loop cycles.  It also updates the wording for
the variable to properly reflect its function.

Change-Id: I0b7f0b3c024e5c66e7b3fb50c711b04556e732fb