Remove log_helper in list_cnas
The list_cnas() method is logging via the pypowervm log_helper
each time a VM is not found. This can generate a lot of
unnecessary logging.  This change removes the logger for
this method only as it's just listing the CNAs for the VM.

Change-Id: Id2f5c114838da293ec2ed3ae38a65c72633f57f6
(cherry picked from commit a4a9a3bb388b2528cf86e0806733cd1d36ad8b9c)