Optimize retry domain
If the REST API has a significant number of conflicting requests come
in, there are times when it can cause a conflict.  These get
encapsulated into errors.  The retry operator allows us to catch a set
of known errors and retry them automatically (ex. etag conflicts).

The retry 'domain' indicates how much of an operation would be retried.
Usually you want to keep this as small as possible.  Retry as few
commands as needed to the API.

The networking-powervm code has a rather large retry domain in the
list_cnas code, and didn't have any retry in the list_lpar_uuids.  This
change set optimizes the retry domains to be small (one API call instead
of many).

The existing test cases cover the refactoring.

Change-Id: I5b758e03dd15f84e65e3df9f88ad1d6d156ddd34