Move CNA Event Handler to Agent Base
As we evaluate bringing new Neutron agents in for the PowerVM platform,
the CNA Event Handler is seen as needed by those agents (to identify
when a new request comes in).

This refactoring moves the Event Handler up to the base class so that
other agents can make use of it.

Change-Id: I2a6891a1332834808ff9dfc493c32d10fc39dd64