Update flake8 ignore rules
This change set removes some of the flake8 ignore rules that the core
neutron project no longer supports.

It also removes any flake8 ignore rules that the core neutron project
does have - but is looking to actively remove.  This will ensure we do
not add exceptions to our project, making it something we need to fix at
a later point in time.

Change-Id: I55d27362f5db487a517177f61adbd6938b37f46b