Report agent mappings
The agent should report the mappings so that the neutron server can
determine if the agent supports a given physical network.  This change
set updates the powervm agents to properly return the physical network.

For compatibility sake (ex. when the neutron server is mitaka but an
agent is still liberty), it will fall back to the original logic.  This
logic will be removed in the newton release of OpenStack.

Change-Id: Ia5b44b0793d25e8c93b9e05b3716961c184c28e5
Closes-Bug: 1537218