Update heal code to ensure device up
The heal code within the networking-powervm project would ensure that
the VLAN and client device was routed out to the network.  However, due
to it calling 'get_device_details', the neutron code was changing the
state back to BUILD.

Given this behavior, it became apparent that the best path forward was
to have the heal code call a full provision request for the client
device.  This actually will no-op very quickly if the VLAN is already on
the client device, but tells Neutron that it is not in fact in a build
state...but rather is now ACTIVE.

This allows for a more robust provisioning scheme and allows the neutron
state to reflect reality.  It also updates any existing ports in the
field that may be affected by this with the next 'heal' cycle.

Change-Id: I02f2c4cd1d63b7a712e50c273e043e6a7ea5a5e1
Closes-Bug: 1535918