Better default value for heal and optimize loop
Currently heal and optimize loop interval is set to 300 seconds
Due to this, it has been observed that CPU utilization is very
high. This is due to frequent port updates. During every loop,
port is updated to DOWN and made UP.

This work will increase this interval to 1800 seconds (30 minutes).

Change-Id: I3ed9abca5534228eb335ae21070cbf73b9f43aed
Closes-Bug: #1571648