Support returning non-mgmt LPAR UUIDs
This change allows for the utils list_lpar_uuids to take in a new
parameter that can exclude the mgmt UUID.  This is an important change
because we may wish to explicitly exclude it to avoid erronerous logging
messages.  This differs from pypowervm's base implementation where it
explicitly includes it (if that explicit include is not set to True, it
does not explicitly exclude it).

Reworks to use the pypowervm get_partitions method where able.

Closes-Bug #1602598

Change-Id: Ic47b4f20c04a7f98517a77cdb31d044a1f8f63e8