Fix bridge mapping change
The related_href was used to parse out the bridge_mappings table.
Unfortunately, *how* you queried for the data identified what the URI
would be.  An earlier change to remove some host_uuid's ended up
malforming the URI and breaking the bridge_mapping logic.

This change fixes how we look at the URIs to see if a SEA is on a given
VIOS.  This should be more reliable moving forward.

Change-Id: I98f76a44db0f61d3e8ed4d3204c1e086bfe00fab
Closes-Bug: #1612047