Enforce limit of VLAN clean ups in each pass
If the system has an excessive amount of VLANs to clean up, the heal and
optimization path can block the requests coming in for an extensive
period of time.  This may eventually lead to VIF plugging timeouts on
the nova component.

This change set will stop doing deletes of VLANs after three passes.  It
logs the remaining VLANs that need to be cleaned up, but does not
actually execute the delete.  It assumes the next pass through will do
the appropriate cleanup.

Change-Id: I97d74531b542177b105ca2625949075b1c8fbf24
Closes-Bug: #1612047