Remove the PVID Looper
With the ability to pass the VLAN over to nova, via the mechanism driver
and vif binding details, we no longer need the PVID Updater.

Previously the function would wait for nova to provision a VIF on VLAN
1.  Then the sea networking agent would detect this new port, update the
SEA to support the VLAN (if needed) and then update the VLAN on the VIF.

Now that we can pass the VLAN to Nova at the VIF create time, that last
step is not needed.

Change-Id: Ifc5bfe8b5e118f02faf0556cc3e67d11a66e2b25