capacity from binding:profile => ml2 conf => None
Previously, the VF capacity for SR-IOV-backed vNICs had to come from the
vif's binding:profile; else it would default to None (allowing the
platform to default it).

With this change, the capacity can also be specified in the ml2
configuration file via the new 'vnic_vf_capacity' option in the AGENT
section.  The binding:profile value takes precedence, followed by the
configuration value.  The configuration default is None if unspecified.

Since binding:profile values will not be supported in the current
release (see Related-Bug), references thereto are removed from the
devref.

Related-Bug: 1615128

Change-Id: I91dc7cf5820bbdd66af80e8f492a80d03c21c02e