Update Hacking Requirement
This updates the hacking requirement to include the off_by_default
checks.

In addition there is a pep8 issue affecting some back ported fixes to
stable/mitaka. This was fixed in master under some new function defect
without anyone being fully aware it was happening. This bug is to
unwedge fix backports to stable/mitaka.

Closes-Bug: 1621292
Change-Id: I9f15c2509d74c81bc366f80d3e08d9bc469e33c4