Time out port update waiting for vif plug
When the SR-IOV agent receives a port update notice from neutron, it
polls the server waiting for the corresponding vNIC to be created before
marking the port as active.  If the vNIC create errors or doesn't
happen, the agent would keep polling for that vNIC indefinitely.

With this change set, if the vNIC doesn't spring into existence within
20 minutes, the agent gives up and stops polling.  (If whatever caused
the vNIC not to be created hasn't already triggered rollback/cleanup of
the corresponding compute process, that process also times out after 20
minutes.)

Closes-Bug: #1629559
Change-Id: I5fd640dfb799c51738aa3613852b4ac69decb31b