Heal and optimize fails to retrieve vifs.
heal_and_optimize flow retrieves a list of vifs
using a method in util module. For SR-IOV agent,
_find_vifs methid is invoked, it is common to CNA
and VNIC objects. The vswitch_id is retrieved and
validated, since vswitch_id is not present in VNIC
object, getattr call fails. Since this method is
common to both CNA and VNIC, a default return value
of None should be included in the list of parameters.

networking_powervm/plugins/ibm/agent/powervm/util.py
_find_vifs method :
 Instead of getattr(vif, 'vswitch_id')
 getattr(vif, 'vswitch_id', None) should be used.

Change-Id: Ifcbf0a9b9d5eb9a326c93cac675b19a2fefca75c
Closes-Bug: #1691584