Move neutron to requirements.txt
Neutron was added to global-requirements [1] which breaks the tox
install of neutron. Neutron is now listed in requirements.txt and has
been removed as a dependency in tox.ini. One side effect is that we
based the policy.json path on the location of the neutron git directory
which is no longer the case. The test code has been updated to reflect
the new location of the policy file.

[1] https://review.openstack.org/#/c/550475/

Change-Id: Ibc42b8a835d1a5e6ce09640f5bdf833c2171a9a9