update translations for queens
Update translations for the Queens release.

Note: this is being cherry-picked from stable/queens to master
instead of the other way around. The translations were actually done
based on queens, so that is where they were originally merged. In
hindsight, they should also be merged in master so that it has the
latest translations as well, so we should have followed the normal
process of merging in master first, but too late for that now.

Change-Id: I03ab7bef6e9826631e45e3cb62c7f58600e3cd81
(cherry picked from commit eb5d4bf7f4cca4b3b4a046985f626c8cfc28449a)