tox needs to pull from stable/queens
Updates the tox.ini to pull from neutron's stable/queens branch
instead of master. This should have been done when we first
branched queens here, but was missed.

Also fixes a UT to use neutron/neutron/tests/etc/policy.json instead
of neutron/etc/policy.json. Local UT passes without this change, but
zuul does not pass. It is not clear why this is required for queens,
but we did have to do this for master recently so at least it is
consistent with that, and it does make sense for a UT to use the
test-specific policy.json.

Change-Id: I864422fbed115871ed3b9de8e1702be9f6a255ae