Uncap eventlet and sync requirements
We will manage the eventlet version using constraints now. See the
thread starting at
http://lists.openstack.org/pipermail/openstack-dev/2018-April/129096.html
for more details.

Meanwhile, it was necessary to address these issues in
check-requirements:

Package u'setuptools' is used in requirements.txt but not in
lower-constraints.txt

Package u'sphinx' is constrained to 1.6.5 which does not match the
minimum version specifier 1.6.2 in test-requirements.txt

Change-Id: Ie0e7b132acab2300e766ae57613a702db21b142f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>