More py3.x support
The Stein supported runtimes proposal [1] has chosen py36 over py35
due to the communities move to testing on Ubuntu 18.04, where py36
is the default runtime. We update tox.ini and setup.cfg accordingly.

Python 3.6 handles dict order differently than in previous python
versions, so to get py36 to work this uses a sorted() call to make
the behavior consistent across python versions for our network
bridge cleanup and adjusts the corresponding test accordingly.

We also add a py37 option to the tox.ini to facilitate future-looking
tests per the resolution on keeping up with Python 3 releases [2].

The py35 option will eventually need to be removed from tox.ini, but
this keeps it for now in order to allow time for folks to transition.

As previously commented, automatic tox envs (pyXX) will use the python
version appropriate to that env and ignore basepython inherited from
[testenv], so we no longer specify a basepython explicitly for these.

[1] https://review.openstack.org/611080
[2] https://review.openstack.org/613145

Change-Id: Iebdc9357a79cae1923109fb5e37fefb8021af18d